fix(mqtt): isolate SDK work and require fresh publication proof
Move producers and downlink execution off the SDK path, preserve bounded FIFO admission and reserve critical capacity in both queues. Correlate probes across reconnects, gate OTA confirmation on current proof, and only revive SDK tasks after verified cleanup. Add Linux ASan/UBSan host coverage, SDK lifecycle pins, CI regressions and bench documentation.

+++ firmware-ESP32-LTE-M/main/oled_menu.c
@@ -94,8 +94,8 @@ static void bal_format(char *out, size_t cap, const arkiv_tlm_balance_t *b)
     const char *hint = b->wei < BAL_WORST_WRITE_WEI ? "\nno gas"
                      : b->wei < BAL_LOW_WEI         ? "\nlow gas" : stale;
     snprintf(out, cap, "BALANCE\n%" PRIu64 ".%06" PRIu64 "\nW3P%s",
-             b->wei / 1000000000000000000ULL,
-             (b->wei % 1000000000000000000ULL) / 1000000000000ULL, hint);
+             (uint64_t)(b->wei / 1000000000000000000ULL),
+             (uint64_t)((b->wei % 1000000000000000000ULL) / 1000000000000ULL), hint);
 }
 /* BAL_FORMAT_END */
